+++ test/integration/test_qt.py
@@ -6,7 +6,7 @@ from . import *
 class TestQt(IntegrationTest):
     def run_executable(self, exe):
         if env.host_platform.genus == 'linux':
-            output = self.assertPopen([exe], env={'DISPLAY': ''},
+            output = self.assertPopen([exe], extra_env={'DISPLAY': ''},
                                       returncode='fail')
             self.assertRegex(output,
                              r'QXcbConnection: Could not connect to display')
@@ -15,8 +15,8 @@ class TestQt(IntegrationTest):
         env_vars = ({} if env.builder('c++').flavor == 'msvc'
                     else {'CPPFLAGS': ('-Wno-inconsistent-missing-override ' +
                                        env.getvar('CPPFLAGS', ''))})
-        super().__init__(os.path.join(examples_dir, '13_qt'), env=env_vars,
-                         *args, **kwargs)
+        super().__init__(os.path.join(examples_dir, '13_qt'),
+                         *args, extra_env=env_vars, **kwargs)
 
     def test_designer(self):
         self.build(executable('qt-designer'))